refactor: export common methods (#21)
* chore: added missing tables and columns to database.md

* docs: limited name and symbol sizes on database design

* refactor: moved common methods to a commons.ts file
