+++ src/txProcessor.ts
@@ -6,10 +6,15 @@
  */
 
 import { APIGatewayProxyResult, SQSEvent } from 'aws-lambda';
-import { ServerlessMysql } from 'serverless-mysql';
 import 'source-map-support/register';
 import hathorLib from '@hathor/wallet-lib';
 
+import {
+  getAddressBalanceMap,
+  getWalletBalanceMap,
+  markLockedOutputs,
+  unlockUtxos,
+} from '@src/commons';
 import {
   addNewAddresses,
   addUtxos,
@@ -20,20 +25,13 @@ import {
   maybeUpdateLatestHeight,
   removeUtxos,
   storeTokenInformation,
-  unlockUtxos as dbUnlockUtxos,
   updateAddressTablesWithTx,
-  updateAddressLockedBalance,
-  updateWalletLockedBalance,
   updateWalletTablesWithTx,
 } from '@src/db';
 import {
-  DecodedOutput,
   StringMap,
   Transaction,
   TokenBalanceMap,
-  TxInput,
-  TxOutput,
-  Utxo,
   Wallet,
 } from '@src/types';
 import { closeDbConnection, getDbConnection, getUnixTimestamp } from '@src/utils';
@@ -143,151 +141,4 @@ const addNewTx = async (tx: Transaction, now: number, blockRewardLock: number) =
   // update wallet_balance and wallet_tx_history tables
   const walletBalanceMap: StringMap<TokenBalanceMap> = getWalletBalanceMap(addressWalletMap, addressBalanceMap);
   await updateWalletTablesWithTx(mysql, txId, tx.timestamp, walletBalanceMap);
-};
-
-/**
- * Mark a transaction's outputs that are locked. Modifies the outputs in place.
- *
- * @remarks
- * The timestamp is used to determine if each output is locked by time. On the other hand, `hasHeightLock`
- * applies to all outputs.
- *
- * The idea is that `hasHeightLock = true` should be used for blocks, whose outputs are locked by
- * height. Timelocks are handled by the `now` parameter.
- *
- * @param outputs - The transaction outputs
- * @param now - Current timestamp
- * @param hasHeightLock - Flag that tells if outputs are locked by height
- */
-export const markLockedOutputs = (outputs: TxOutput[], now: number, hasHeightLock = false): void => {
-  for (const output of outputs) {
-    output.locked = false;
-    if (hasHeightLock || output.decoded.timelock > now) {
-      output.locked = true;
-    }
-  }
-};
-
-/**
- * Get the map of token balances for each address in the transaction inputs and outputs.
- *
- * @example
- * Return map has this format:
- * ```
- * {
- *   address1: {token1: balance1, token2: balance2},
- *   address2: {token1: balance3}
- * }
- * ```
- *
- * @param inputs - The transaction inputs
- * @param outputs - The transaction outputs
- * @returns A map of addresses and its token balances
- */
-export const getAddressBalanceMap = (
-  inputs: TxInput[],
-  outputs: TxOutput[],
-): StringMap<TokenBalanceMap> => {
-  const addressBalanceMap = {};
-
-  for (const input of inputs) {
-    const address = input.decoded.address;
-
-    // get the TokenBalanceMap from this input
-    const tokenBalanceMap = TokenBalanceMap.fromTxInput(input);
-    // merge it with existing TokenBalanceMap for the address
-    addressBalanceMap[address] = TokenBalanceMap.merge(addressBalanceMap[address], tokenBalanceMap);
-  }
-
-  for (const output of outputs) {
-    const address = output.decoded.address;
-
-    // get the TokenBalanceMap from this output
-    const tokenBalanceMap = TokenBalanceMap.fromTxOutput(output);
-    // merge it with existing TokenBalanceMap for the address
-    addressBalanceMap[address] = TokenBalanceMap.merge(addressBalanceMap[address], tokenBalanceMap);
-  }
-
-  return addressBalanceMap;
-};
-
-/**
- * Get the map of token balances for each wallet.
- *
- * @remarks
- * Different addresses can belong to the same wallet, so this function merges their
- * token balances.
- *
- * @example
- * Return map has this format:
- * ```
- * {
- *   wallet1: {token1: balance1, token2: balance2},
- *   wallet2: {token1: balance3}
- * }
- * ```
- *
- * @param addressWalletMap - Map of addresses and corresponding wallets
- * @param addressBalanceMap - Map of addresses and corresponding token balances
- * @returns A map of wallet ids and its token balances
- */
-export const getWalletBalanceMap = (
-  addressWalletMap: StringMap<Wallet>,
-  addressBalanceMap: StringMap<TokenBalanceMap>,
-): StringMap<TokenBalanceMap> => {
-  const walletBalanceMap = {};
-  for (const [address, balanceMap] of Object.entries(addressBalanceMap)) {
-    const wallet = addressWalletMap[address];
-    const walletId = wallet && wallet.walletId;
-
-    // if this address is not from a started wallet, ignore
-    if (!walletId) continue;
-
-    walletBalanceMap[walletId] = TokenBalanceMap.merge(walletBalanceMap[walletId], balanceMap);
-  }
-  return walletBalanceMap;
-};
-
-/**
- * Update the unlocked/locked balances for addresses and wallets connected to the given UTXOs.
- *
- * @param _mysql - Database connection
- * @param utxos - List of UTXOs that are unlocked by height
- * @param updateTimelocks - If this update is triggered by a timelock expiring, update the next lock expiration
- */
-export const unlockUtxos = async (_mysql: ServerlessMysql, utxos: Utxo[], updateTimelocks: boolean): Promise<void> => {
-  if (utxos.length === 0) return;
-
-  const outputs: TxOutput[] = utxos.map((utxo) => {
-    const decoded: DecodedOutput = {
-      type: 'P2PKH',
-      address: utxo.address,
-      timelock: utxo.timelock,
-    };
-    return {
-      value: utxo.authorities > 0 ? utxo.authorities : utxo.value,
-      token: utxo.tokenId,
-      decoded,
-      locked: false,
-      // set authority bit if necessary
-      token_data: utxo.authorities > 0 ? hathorLib.constants.TOKEN_AUTHORITY_MASK : 0,
-      // we don't care about spent_by and script
-      spent_by: null,
-      script: '',
-    };
-  });
-
-  // mark as unlocked in database (this just changes the 'locked' flag)
-  await dbUnlockUtxos(_mysql, utxos);
-
-  const addressBalanceMap: StringMap<TokenBalanceMap> = getAddressBalanceMap([], outputs);
-  // update address_balance table
-  await updateAddressLockedBalance(_mysql, addressBalanceMap, updateTimelocks);
-
-  // check if addresses belong to any started wallet
-  const addressWalletMap: StringMap<Wallet> = await getAddressWalletInfo(_mysql, Object.keys(addressBalanceMap));
-
-  // update wallet_balance table
-  const walletBalanceMap: StringMap<TokenBalanceMap> = getWalletBalanceMap(addressWalletMap, addressBalanceMap);
-  await updateWalletLockedBalance(_mysql, walletBalanceMap, updateTimelocks);
 };
