+++ tests/txProcessor.test.ts
@@ -2,7 +2,6 @@ import eventTemplate from '@events/eventTemplate.json';
 import tokenCreationTx from '@events/tokenCreationTx.json';
 import { getLatestHeight, getTokenInformation } from '@src/db';
 import * as txProcessor from '@src/txProcessor';
-import { Authorities, Balance, TokenBalanceMap, Utxo } from '@src/types';
 import { closeDbConnection, getDbConnection, isAuthority } from '@src/utils';
 import {
   XPUBKEY,
@@ -40,256 +39,6 @@ afterAll(async () => {
   await closeDbConnection(mysql);
   // restore old env
   process.env = OLD_ENV;
-});
-
-test('markLockedOutputs and getAddressBalanceMap', () => {
-  expect.hasAssertions();
-  const evt = JSON.parse(JSON.stringify(eventTemplate));
-  const tx = evt.Records[0].body;
-  const now = 20000;
-  tx.tx_id = 'txId1';
-  tx.timestamp = 0;
-  tx.inputs = [
-    createInput(10, 'address1', 'inputTx', 0, 'token1'),
-    createInput(5, 'address1', 'inputTx', 0, 'token1'),
-    createInput(7, 'address1', 'inputTx', 1, 'token2'),
-    createInput(3, 'address2', 'inputTx', 2, 'token1'),
-  ];
-  tx.outputs = [
-    createOutput(5, 'address1', 'token1'),
-    createOutput(2, 'address1', 'token3'),
-    createOutput(11, 'address2', 'token1'),
-  ];
-  const map1 = new TokenBalanceMap();
-  map1.set('token1', new Balance(-10, 0));
-  map1.set('token2', new Balance(-7, 0));
-  map1.set('token3', new Balance(2, 0));
-  const map2 = new TokenBalanceMap();
-  map2.set('token1', new Balance(8, 0));
-  const expectedAddrMap = {
-    address1: map1,
-    address2: map2,
-  };
-
-  txProcessor.markLockedOutputs(tx.outputs, now, false);
-  for (const output of tx.outputs) {
-    expect(output.locked).toBe(false);
-  }
-
-  const addrMap = txProcessor.getAddressBalanceMap(tx.inputs, tx.outputs);
-  expect(addrMap).toStrictEqual(expectedAddrMap);
-
-  // update tx to contain outputs with timelock
-  tx.outputs[0].decoded.timelock = now - 1;   // won't be locked
-  tx.outputs[1].decoded.timelock = now;       // won't be locked
-  tx.outputs[2].decoded.timelock = now + 1;   // locked
-
-  // should mark the corresponding output as locked
-  txProcessor.markLockedOutputs(tx.outputs, now, false);
-  expect(tx.outputs[0].locked).toBe(false);
-  expect(tx.outputs[1].locked).toBe(false);
-  expect(tx.outputs[2].locked).toBe(true);
-
-  // check balance
-  map2.set('token1', new Balance(-3, 11, now + 1));
-  const addrMap2 = txProcessor.getAddressBalanceMap(tx.inputs, tx.outputs);
-  expect(addrMap2).toStrictEqual(expectedAddrMap);
-
-  // a block will have its rewards locked, even with no timelock
-  tx.inputs = [];
-  tx.outputs = [
-    createOutput(100, 'address1', 'token1'),
-  ];
-  txProcessor.markLockedOutputs(tx.outputs, now, true);
-  for (const output of tx.outputs) {
-    expect(output.locked).toBe(true);
-  }
-  const addrMap3 = txProcessor.getAddressBalanceMap(tx.inputs, tx.outputs);
-  const map3 = new TokenBalanceMap();
-  map3.set('token1', new Balance(0, 100));
-  const expectedAddrMap2 = {
-    address1: map3,
-  };
-  expect(addrMap3).toStrictEqual(expectedAddrMap2);
-
-  // tx with authorities
-  tx.inputs = [
-    createInput(0b01, 'address1', 'inputTx', 0, 'token1', null, 129),
-    createInput(0b10, 'address1', 'inputTx', 1, 'token2', null, 129),
-  ];
-  tx.outputs = [
-    createOutput(0b01, 'address1', 'token1', null, false, 129),
-    createOutput(0b10, 'address1', 'token2', 1000, true, 129),
-  ];
-  const map4 = new TokenBalanceMap();
-  map4.set('token1', new Balance(0, 0, null));
-  map4.set('token2', new Balance(0, 0, 1000, new Authorities([-1, 0]), new Authorities([1, 0])));
-  const expectedAddrMap4 = {
-    address1: map4,
-  };
-  const addrMap4 = txProcessor.getAddressBalanceMap(tx.inputs, tx.outputs);
-  expect(addrMap4).toStrictEqual(expectedAddrMap4);
-});
-
-test('getWalletBalanceMap', () => {
-  expect.hasAssertions();
-  const mapAddress1 = new TokenBalanceMap();
-  mapAddress1.set('token1', new Balance(-10, 0));
-  mapAddress1.set('token2', new Balance(-7, 0));
-  mapAddress1.set('token3', new Balance(2, 0));
-  const mapAddress2 = new TokenBalanceMap();
-  mapAddress2.set('token1', new Balance(8, 0));
-  const mapAddress3 = new TokenBalanceMap();
-  mapAddress3.set('token2', new Balance(2, 0));
-  mapAddress3.set('token3', new Balance(6, 0));
-  const mapAddress4 = new TokenBalanceMap();
-  mapAddress4.set('token1', new Balance(2, 0));
-  mapAddress4.set('token2', new Balance(9, 1, 500));
-  const mapAddress5 = new TokenBalanceMap();
-  mapAddress5.set('token1', new Balance(11, 0));
-  const addressBalanceMap = {
-    address1: mapAddress1,
-    address2: mapAddress2,
-    address3: mapAddress3,
-    address4: mapAddress4,
-    address5: mapAddress5,    // doesn't belong to any started wallet
-  };
-  const walletAddressMap = {
-    address1: { walletId: 'wallet1', xpubkey: 'xpubkey1', maxGap: 5 },
-    address2: { walletId: 'wallet1', xpubkey: 'xpubkey1', maxGap: 5 },
-    address4: { walletId: 'wallet1', xpubkey: 'xpubkey1', maxGap: 5 },
-    address3: { walletId: 'wallet2', xpubkey: 'xpubkey2', maxGap: 5 },
-  };
-  const mapWallet1 = new TokenBalanceMap();
-  mapWallet1.set('token1', new Balance(0, 0));
-  mapWallet1.set('token2', new Balance(2, 1, 500));
-  mapWallet1.set('token3', new Balance(2, 0));
-  const mapWallet2 = new TokenBalanceMap();
-  mapWallet2.set('token2', new Balance(2, 0));
-  mapWallet2.set('token3', new Balance(6, 0));
-  const expectedWalletBalanceMap = {
-    wallet1: mapWallet1,
-    wallet2: mapWallet2,
-  };
-  const walletBalanceMap = txProcessor.getWalletBalanceMap(walletAddressMap, addressBalanceMap);
-  expect(walletBalanceMap).toStrictEqual(expectedWalletBalanceMap);
-
-  // if walletAddressMap is empty, should also return an empty object
-  const walletBalanceMap2 = txProcessor.getWalletBalanceMap({}, addressBalanceMap);
-  expect(walletBalanceMap2).toStrictEqual({});
-});
-
-test('unlockUtxos', async () => {
-  expect.hasAssertions();
-  const reward = 6400;
-  const txId1 = 'txId1';
-  const txId2 = 'txId2';
-  const txId3 = 'txId3';
-  const txId4 = 'txId4';
-  const txId5 = 'txId5';
-  const token = 'tokenId';
-  const addr = 'address';
-  const walletId = 'walletId';
-  const now = 1000;
-  await addToUtxoTable(mysql, [
-    // blocks with heightlock
-    [txId1, 0, token, addr, reward, 0, null, 3, true],
-    [txId2, 0, token, addr, reward, 0, null, 4, true],
-    // some transactions with timelock
-    [txId3, 0, token, addr, 2500, 0, now, null, true],
-    [txId4, 0, token, addr, 2500, 0, now * 2, null, true],
-    [txId5, 0, token, addr, 0, 0b10, now * 3, null, true],
-  ]);
-
-  await addToWalletTable(mysql, [
-    [walletId, 'xpub', 'ready', 10, now, now + 1],
-  ]);
-
-  await addToAddressTable(mysql, [
-    { address: addr, index: 0, walletId, transactions: 1 },
-  ]);
-
-  await addToAddressBalanceTable(mysql, [
-    [addr, token, 0, 2 * reward + 5000, now, 5, 0, 0b10],
-  ]);
-
-  await addToWalletBalanceTable(mysql, [{
-    walletId,
-    tokenId: token,
-    unlockedBalance: 0,
-    lockedBalance: 2 * reward + 5000,
-    unlockedAuthorities: 0,
-    lockedAuthorities: 0b10,
-    timelockExpires: now,
-    transactions: 5,
-  }]);
-
-  const utxo: Utxo = {
-    txId: txId1,
-    index: 0,
-    tokenId: token,
-    address: addr,
-    value: reward,
-    authorities: 0,
-    timelock: null,
-    heightlock: 3,
-    locked: true,
-  };
-
-  // unlock txId1
-  await txProcessor.unlockUtxos(mysql, [utxo], false);
-  await expect(
-    checkUtxoTable(mysql, 5, txId1, 0, utxo.tokenId, utxo.address, utxo.value, 0, utxo.timelock, utxo.heightlock, false),
-  ).resolves.toBe(true);
-  await expect(checkAddressBalanceTable(mysql, 1, addr, token, reward, reward + 5000, now, 5, 0, 0b10)).resolves.toBe(true);
-  await expect(checkWalletBalanceTable(mysql, 1, walletId, token, reward, reward + 5000, now, 5, 0, 0b10)).resolves.toBe(true);
-
-  // unlock txId2
-  utxo.txId = txId2;
-  utxo.heightlock = 4;
-  await txProcessor.unlockUtxos(mysql, [utxo], false);
-  await expect(
-    checkUtxoTable(mysql, 5, txId2, 0, utxo.tokenId, utxo.address, utxo.value, 0, utxo.timelock, utxo.heightlock, false),
-  ).resolves.toBe(true);
-  await expect(checkAddressBalanceTable(mysql, 1, addr, token, 2 * reward, 5000, now, 5, 0, 0b10)).resolves.toBe(true);
-  await expect(checkWalletBalanceTable(mysql, 1, walletId, token, 2 * reward, 5000, now, 5, 0, 0b10)).resolves.toBe(true);
-
-  // unlock txId3, txId4 is still locked
-  utxo.txId = txId3;
-  utxo.value = 2500;
-  utxo.timelock = now;
-  utxo.heightlock = null;
-  await txProcessor.unlockUtxos(mysql, [utxo], true);
-  await expect(
-    checkUtxoTable(mysql, 5, txId3, 0, utxo.tokenId, utxo.address, utxo.value, 0, utxo.timelock, utxo.heightlock, false),
-  ).resolves.toBe(true);
-  await expect(checkAddressBalanceTable(mysql, 1, addr, token, 2 * reward + 2500, 2500, 2 * now, 5, 0, 0b10)).resolves.toBe(true);
-  await expect(checkWalletBalanceTable(mysql, 1, walletId, token, 2 * reward + 2500, 2500, 2 * now, 5, 0, 0b10)).resolves.toBe(true);
-
-  // unlock txId4
-  utxo.txId = txId4;
-  utxo.value = 2500;
-  utxo.timelock = now * 2;
-  utxo.heightlock = null;
-  await txProcessor.unlockUtxos(mysql, [utxo], true);
-  await expect(
-    checkUtxoTable(mysql, 5, txId4, 0, utxo.tokenId, utxo.address, utxo.value, 0, utxo.timelock, utxo.heightlock, false),
-  ).resolves.toBe(true);
-  await expect(checkAddressBalanceTable(mysql, 1, addr, token, 2 * reward + 5000, 0, 3 * now, 5, 0, 0b10)).resolves.toBe(true);
-  await expect(checkWalletBalanceTable(mysql, 1, walletId, token, 2 * reward + 5000, 0, 3 * now, 5, 0, 0b10)).resolves.toBe(true);
-
-  // unlock txId5
-  utxo.txId = txId5;
-  utxo.value = 0;
-  utxo.authorities = 0b10;
-  utxo.timelock = now * 3;
-  utxo.heightlock = null;
-  await txProcessor.unlockUtxos(mysql, [utxo], true);
-  await expect(
-    checkUtxoTable(mysql, 5, txId5, 0, utxo.tokenId, utxo.address, utxo.value, utxo.authorities, utxo.timelock, utxo.heightlock, false),
-  ).resolves.toBe(true);
-  await expect(checkAddressBalanceTable(mysql, 1, addr, token, 2 * reward + 5000, 0, null, 5, 0b10, 0)).resolves.toBe(true);
-  await expect(checkWalletBalanceTable(mysql, 1, walletId, token, 2 * reward + 5000, 0, null, 5, 0b10, 0)).resolves.toBe(true);
 });
 
 /*
